% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg.loglik.WeibExp.R
\name{neg.loglik.WeibExp}
\alias{neg.loglik.WeibExp}
\title{Negative Log-Likelihood for Weibull-Exponential Distribution}
\usage{
neg.loglik.WeibExp(param, changeP, time, event)
}
\arguments{
\item{param}{Shape and scale parameter for Weibull distribution.}

\item{changeP}{Changepoint.}

\item{time}{Vector of survival times.}

\item{event}{Vector indicating censoring status; 0 = alive (censored), 1 = dead
(uncensored).}
}
\value{
Value of the negative log-likelihood.
}
\description{
Negative Log-Likelihood for Weibull-Exponential Distribution
}
